Added player manager, added bridge
