--- Bat_Brewery/Assets/Environment/ObjectFading.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: fdd1d1ddcc469624bad0761e7aaf40e6
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
