--- Bat_Brewery/Assets/Environment/ObjectRegionCulling.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 05c49450929da524d9ff54ff412e44f0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
