--- Bat_Brewery/Assets/Environment/ObjectRenderStacking.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f5557e79405b0d646b6037bb67888eaa
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
