--- Bat_Brewery/Assets/Environment/ObjectSortingOrderer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bead5ae545b173b488fdb8a045e22a63
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
