+++ Bat_Brewery/Assets/Environment/Objects/ObjectObstructionFading.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 698b3d089fcb6b04ab94ce2b1b25fd7b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
