+++ Bat_Brewery/Assets/Environment/Objects/ObjectQuestChange.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b9fbfe6a8cd92ba4cb8100b7e2ed858c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
