+++ Bat_Brewery/Assets/Managers.prefab
@@ -33,6 +33,7 @@ Transform:
   - {fileID: 4216814360938530774}
   - {fileID: 5155925886494411525}
   - {fileID: 7611517790273625848}
+  - {fileID: 6204955785852836041}
   m_Father: {fileID: 0}
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1001 &1453553199332994771
@@ -108,6 +109,68 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 5ba8f45dccf07c04e97c77ae0eb197ab, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!1001 &1548233377753966872
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 7265043287754539996}
+    m_Modifications:
+    - target: {fileID: 3613688364231317915, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_Name
+      value: FootstepController
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+--- !u!4 &6204955785852836041 stripped
+Transform:
+  m_CorrespondingSourceObject: {fileID: 4854880830632161745, guid: 8ea48549e10124f428610c2ae996c211, type: 3}
+  m_PrefabInstance: {fileID: 1548233377753966872}
+  m_PrefabAsset: {fileID: 0}
 --- !u!1001 &4213081484328187103
 PrefabInstance:
   m_ObjectHideFlags: 0
