--- Bat_Brewery/Assets/Player/Footsteps/FootstepController.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 8ea48549e10124f428610c2ae996c211
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
