+++ Bat_Brewery/Assets/Player/Footsteps/FootstepManager.prefab
@@ -1,0 +1,58 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1 &3613688364231317915
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 4854880830632161745}
+  - component: {fileID: 2970915126502391724}
+  m_Layer: 0
+  m_Name: FootstepManager
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &4854880830632161745
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3613688364231317915}
+  serializedVersion: 2
+  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 0}
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &2970915126502391724
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 3613688364231317915}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 190a6d2ba89684a4f8a967a096906f8e, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  playerFeet: {fileID: 0}
+  footstepPrefab:
+  - {fileID: 6638812221064593384, guid: 1c6c4db6e8bd06841bf5c1aa84fd3bac, type: 3}
+  - {fileID: 6638812221064593384, guid: 4ca6067146fd28f44bf7fd64154ad929, type: 3}
+  - {fileID: 6638812221064593384, guid: 37f3c3b41f290ad4c95089780aa92793, type: 3}
+  - {fileID: 6638812221064593384, guid: a9d6f4edc4ffcf44f892485d67b415e7, type: 3}
+  - {fileID: 6638812221064593384, guid: 8b09a66a92198fa4681ee8b54d1e440e, type: 3}
+  - {fileID: 6638812221064593384, guid: cefac5f03df75fa498fcf4bc8feeccb3, type: 3}
+  maximumFootsteps: 6
+  footstepDistance: 2
+  playerNoFootstepRadius: 2.5
+  playerFadeFootstepsRadius: 3.5
