+++ Bat_Brewery/Assets/Player/Footsteps/FootstepManager.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 8ea48549e10124f428610c2ae996c211
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
