+++ Bat_Brewery/Assets/Player/Vision/PlayerVisionController.cs
@@ -4,8 +4,6 @@ using UnityEngine;
 
 public class PlayerVisionController : MonoBehaviour
 {
-    [SerializeField] Transform playerTransform;
     [SerializeField] public float maxViewDistance;
     [SerializeField] public float maxClearDistance;
-    public Transform GetPlayerTransform(){return playerTransform;}
 }
